Allow setting random delays globally
Just like fixed delays, randomly distributed delays can be set for all
stubs.

It's a bit messier than I'd like, but the existing global fixed delay is
a bit difficult to expand.

I've tried to keep it consistent with stubbing in that the delays are
independent of each other and can be combined.