Add response templating for proxyBaseUrl
This extends the response templating extension to work with proxy URLs. This is very useful when
using the proxy/intercept pattern with multiple different API providers, especially when the
endpoints aren't known ahead of time. The browser proxy mode is perhaps better suited to this use
case, but it's not an option for HTTPS-enabled APIs due to the lack of SSL support.