Add getOptions() to Admin interface
For the new recording API endpoint, we'll want to be able to extract response
bodies to a file and load stub mapping transformer extensions. This means
having access to `Options.filesSource()` so it can write out the extracted
bodies, and `Options.extensionsOfType()` for loading the transformer
extensions.

I'm not sure if `WireMockServer.getOptions()` and `WireMockApp.getOptions()`
are really worth testing, since they're simple getters. I wrote a test for
former since there's already a test class for it.