Fix JsonFileMappingsSource.remove() for subdirectories
The "JsonFileMappingsSource.loadingMappingsInfo()" method looks for JSON
files in the mappings directory recursively, which is useful since it
allows organizing stubs in a directory hierachy. However, the "remove()"
and "removeAll()" methods will not remove stubs in a subdirectory,
because the path components are stripped out of the filename.

This was introduced in 7f93b5eb996722c1919a6e1e627d550f7929ac39, which
changed the "getFileName()" to strip out all path compontents. I updated
that method to only do so for non-file URLs.