Add RequestPatternBuilder.like() and but() methods
Sometimes, when writing a StubMappingTransformer extension, you want to
change a single field in the RequestPattern (e.g. the URL). This is
difficult because RequestPattern is immutable and RequestPatternBuilder
can't build off an existing object the way ResponseDefinitionBuilder
can.

This adds two new methods to RequestPatternBuilder, like() and but(),
that do the same thing as the corresponding methods in
ResponseDefinitionBuilder.