Updates and fixes for RAML API spec
This makes some updates to the RAML specs, mainly to flesh out the
recording API. I also made a few fixes so that running
`api-spec-converter -f raml -t openapi_3 tmp/raml/wiremock-admin-api.raml`
generates a valid OpenAPI 3.0 (aka Swagger 3) spec file (requires
upgrading `api-spec-converter`). It's not valid Swagger 2, but that was
already the case due to the use of `oneOf` for `delayDistribution`,
which is not supported in Swagger 2.

The latest version of Swagger UI supports OpenAPI 3, so I think it'd be
a good idea to upgrade, but I didn't do that here.