Update to Swagger UI 3
This updates Swagger UI to the latest version, since the current
version doesn't support OpenAPI 3.0. Although I could've extracted the
new version in place, I figured it'd be better to handle it in
package.json by adding `swagger-ui-dist` as a dependency and copying it
over the necessary files during build time. This will make upgrades
easier in the future.