Add independent OpenAPI 3 spec
This adds an OpenAPI 3.0 spec and uses it to generate the swagger.json
file, instead of converting the RAML spec to Swagger 2.0 with
`api-spec-converter`. OpenAPI is more flexible and solves conversion
issues involving the `$ref` elements.  This new spec is much more
complete than the RAML spec, which will be removed entirely in a
subsequent commit.