Disallow HTTP/2 when proxying HTTPS
If the proxy CONNECT request is made over HTTPS, and the actual content request
is made using HTTP/2 tunneled over HTTPS, and an exception is thrown,
the server blocks for 30 seconds before flushing the response.

To fix this, force HTTP/1.1 over TLS when tunneling HTTPS.

This also means the HTTP connector does not need the alpn & h2 connection
factories as it will not use them, so we can revert to something like the code
before we added HTTPS forward proxying.

Unfortunately it has proven too hard to write a test to demonstrate the bug; it
requires an HTTP client capable of doing ALPN & HTTP/2, which will only offer
HTTP/1.1 in the ALPN negotiation when using HTTPS for the initial CONNECT
request but will then offer both HTTP/1.1 and HTTP/2 for the actual request
(this is how curl 7.64.1 behaves!). Neither Apache HTTP 4, 5, 5 Async, OkHttp,
nor the Jetty client could do this. It might be possible to write one using
Netty, but it would be hard and time consuming.