Generate a certificate per HTTPS proxied target
If the keystore passed to WireMock includes a certificate & private key that
can be used as a certificate authority, it will be used as the root certificate
to sign certificates generated per proxied host.

This should allow the user to add just that root certificate to their trusted
root certificates, and then be able to proxy via WireMock without security
problems.

The security of doing so depends on the security of the private key passed to
WireMock.