Enable support for this codebase's to write JUnit5 tests
JUnit5 is arguably a better development experience, and we should allow
contributors to start writing unit tests using it, instead of JUnit4,
where possible.

However, until we've migrated the codebase between versions, we can use
the vintage-engine to run both JUnit4 and JUnit5 tests at the same time!

This also includes the platform-launcher, which will allow us to launch
the tests from an IDE.

As JUnit4's dependency is brough in transitively from
`junit-vintage-engine`, we can remove the explicit dependency on JUnit4.