Allow standalone runner to fetch mappings from classpath
To allow consumers to shade their standalone JAR with the stubs they
want to produce, for instance to allow there to be a single artefact
with all their stubs and code, we should allow this being specified to
be part of the classpath.

This introduces a new flag, `--load-resources-from-classpath`, to allow
configuring both the `__files` and `mappings` directories to be
retrieved from the classpath.

Closes #1502.