Enforce consistent code style with Spotless
As a first step towards having consistent, enforced code style across
the project, to simplify diffs, and to reduce the effort by
contributors, we can use Spotless.

By default, we'll enable Java, Gradle and JSON configurations, and have
the option of adding other source types in the future.

We need to exclude the `sample.json` file as it's not valid JSON,
otherwise Spotless fails to run.

This also documents how to run Spotless locally, and documents how to
have it running automagically on compiles via https://www.jvt.me/posts/2020/05/15/gradle-spotless/

We make sure this is applied to `allprojects` (which requires the
`apply`), so everything is consistent.

Closes #1559.