Return an Bad Request when invalid `since` parameter is provided
As noted in #1595, we currently cause an Internal Server Error when an
invalid date is provided to the admin `/requests` API.

Instead of using the runtime exception `IllegalArgumentException`, which
is less nice to try-catch, we can take advantage of the
`InvalidInputException`, containing a `validation` error, which can be
returned nicely as a JSON response.

As this is using a JUnit4 test, we need to make sure that ArchUnit
allows it as a valid use case, until `AcceptanceTestBase` is migrated to
JUnit Jupiter.

Closes #1595.