Allow HTTPS browser proxying on J17 as a process
Under Java 13+ (maybe 12+?) access to the `sun.security.x509` classes is
prohibited unless explicitly allowed. This allows it when running
WireMock as a standalone process. We need them for browser proxying of
HTTPS traffic.

Note it is not transitive - if you include WireMock as a library in
another process you will need to explitly allow access by adding this
entry to the main jar's manifest or running with
`--add-exports=java.base/sun.security.x509=ALL-UNNAMED`.