Amend Gradle configuration to be a Library
To simplify the Gradle 7 upgrade, due to the way that the deprecated
`maven` plugin is now removed, we can prefactor by replacing any
configuration around the generating/publishing of the POMs with default
Gradle configuration.

As we want to produce transitive dependencies, we can now do this more
easily with `api` dependencies.