Manage JAR / POM configurations through Gradle
As part of a prefactor for the Gradle 7 upgrade, we can simplify the
creation of our POMs, as there's no longer the `pom` functionality
through the now removed `maven` plugin.

Because the `jar` and `shadowJar` end up with the same classifier `''`,
and are part of the same project, Gralde can't decide which one to use.

To avoid this, we let it use the `-all` classifier, by default, allowing
the thin package to be generated correctly, and then have the standalone
JAR override the classifier at publish time.

This also allows us to produce all of the artifacts for the project more
easily, and in the future, we can produce the correct POM for consumers,
with transitive dependencies, through the `from components.java`.

Unfortunately, this does lead to the standalone JAR being added onto the
project output for the non-standalone project, as it's technically an
artifact of this project.

Consumers of the standalone dependency expect our `sources`, `javadoc`
and `test` JAR to be available to them, so we need to make sure that
they're present in the `standalone` project's list of `artifact`s.

Finally, we make sure that `assemble` waits for `jar` and `shadowJar`,
instead of them being explicitly bound for publishing tasks.