Enforce license headers with Spotless
Previously we've got the `add-copyright-headers` task to retroactively
add license headers, but fortunately, Spotless can do this for us too,
so we can remove our code and apply it as part of a Spotless run.

To simplify the configuration in our `build.gradle`, we can put this
into a separate file that can be edited on its own.

To ensure this stays up-to-date, we need to add `ratchetFrom` to make
sure that this updates against our main branch[0].

We need to make sure that we've updated GitHub Actions to pull the full
Git history - as per a suggestion on
https://github.com/diffplug/spotless/issues/919#issuecomment-904957487
- so we can correctly ratchet as part of our CI.

[0]: https://www.jvt.me/posts/2021/03/31/git-origin-head/