Issue 1989: Support Content-Encoding: deflate
This change adds support for Content-Encoding: deflate, and fixes a bug
so that if an unrecognized Content-Encoding type is encountered we pass
it through instead of removing it and making recorded stubs unusable.

Prior to these changes, if a Content-Encoding header is present, we remove
it.  However, we only decompress the body if the Content-Encoding is gzip.
If we're going to remove the header, we should also decode the content,
and vice-versa.

Also, added .gitattributes so that when spotlessApply is used on Windows
line endings are not changed to CRLF.