ConsoleNotifyingWiremockNetworkTrafficListener accepts custom encoding
Allows specification of a custom `Charset` for on
`ConsoleNotifyingWiremockNetworkTrafficListener` allowing flexibility
for service mocks that include non UTF-8 characters in their responses.

Additionally:
- Adds test coverage for `ConsoleNotifyingWiremockNetworkTrafficListener`.
- Changes behavior of `incoming` and `outgoing` so that if a ByteBuffer
with a Charset other than the one specified by
`ConsoleNotifyingWiremockNetworkTrafficListener` is provided, a succinct
message is logged. This replaces the previous behavior that dumped the
entire CharacterCodingException. This was extremely noisy and did not
provide useful additional diagnostics.

Fixes #1541