Fix json string schema rejecting numbers
When a path param, header value, query string etc. matcher is
`matchesJsonSchema`, and the schema requires the value to be of type
string, then at present it will be rejected if it has value `123` or
`true` or `false` because it will be considered a number or a boolean.

As these are also possible strings, try validating them as text as well.