Add a custom number comparator to the Json Diff to trim trailing zeros
Now we have set our ObjectMapper to not normalise BigDecimal values, we
have an issue with the json Diff not seeing numbers like `1.2` and
`1.2000` as the same values. This change adds a new Normalised Number
Comparator to the Diff so we can continue to compare those values in
the same way as we used to.