Add sub events to the webhook success and failure scenarios
Sub events are now added when we make a webhook request on success or
failure scenarios.  The sub event message is at INFO for success and
ERROR for any error cases.  The text of the sub event is the same as
the message for the notifier with the exception message appended for
any error scenario.