Mark API that does not need to be public deprecated
As far as we can see the `SnapshotStubMappingPostProcessor` and
`SnapshotStubMappingTransformerRunner` do not need to be public API -
they are instantiated explicitly by `Recorder` and cannot be injected
anywhere.

We want to change their API in v4, so reducing the visibility first.

Likewise, we see no benefit to `Recorder.getStubMappingPostProcessor`
and `Recorder.serveEventsToStubMappings` being public, they are only
called by `takeSnapshot`.