When using a dynamic port always bind to loopback
By default, a wiremock instance (including the mock host) binds to
address 0.0.0.0. When binding to a random port, that port is only unique
for bind address 0.0.0.0; another application may be listening on the
same TCP/IP port for a more specific bind address (like localhost). If
so, and you request that port as localhost, you will talk to the other
application, and your test will fail.

By binding explicitly to the IP address of the loopback adapter this
problem is avoided.

Assumption here is that you would always want to do this... whether this
is a correct assumption...