Handle unsupported charsets
`ContentTypeHeader.charset()` is a convenience method that does not
return `null` or `Optional.absent()` when the header is absent, instead
defaulting to UTF-8; instead you can use
`ContentTypeHeader.encodingPart()` to get precise information about
whether it is absent and what it is.

As such it should not throw an exception for an unsupported charset but
fall back on the default, UTF-8.