Remove standaloneOnly
slf4j-nop is packaged in slf4j-api now, so no need to have it as a
separate dependency, in fact with the package relocation it breaks the
way SLF4J tries to notice that there are multiple implementations on the
classpath.

Means we do not need the standaloneOnly configuration at all.